pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "bcgov_arches_common"
readme = "README.md"
authors = []
license = {text = "GNU AGPL3"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
]
requires-python = ">=3.10"
dependencies = [
    "arches==7.6.4",
    #"arches @ git+https://github.com/bcgov/arches.git@stable/7.6.3_bcgov_11578_11679_11689",
]
version = "1.1.0-b0"

[project.optional-dependencies]
dev = [
    "livereload",
    "sst",
    "coverage",
    "django-silk==5.1.0",
    "pre-commit==3.8.0",
    "black==24.4.2",
]

[tool.setuptools]
packages = ["bcgov_arches_common"]
